tests in process... business layer as well... now coding BasketManager...

+++ src/main/java/com/workfront/internship/business/CategoryManagerImpl.java
@@ -23,7 +23,8 @@ public class CategoryManagerImpl implements CategoryManager {
     }
 
     public Category getCategoryByID(int id){
-
+        if(id <=0)
+            throw new RuntimeException("not a valid categoryId");
         Category category = categoryDao.getCategoryByID(id);
         return category;
 
@@ -34,21 +35,19 @@ public class CategoryManagerImpl implements CategoryManager {
         int index = categoryDao.insertCategory(category);
         return index;
     }
-    public void deleteCategoryByID(int id){
-
+    public void deleteCategory(int id){
+        if(id <=0)
+            throw new RuntimeException("not a valid categoryId");
         categoryDao.deleteCategoryByID(id);
 
     }
     public void updateCategory(Category category){
-
+        if(!validateCategory(category))
+            throw new RuntimeException("not a valid category");
         categoryDao.updateCategory(category);
 
     }
-    public void deleteAllCategories(){
 
-        categoryDao.deleteAllCategories();
-
-    }
     public List<Category> getAllCategories(){
 
         List<Category> categories = categoryDao.getAllCategories();
